And now they have health.

+++ stache/source/interface/thinker.d
@@ -28,7 +28,7 @@ interface ISheeple
 	@property bool CanAttack();
 	@property bool CanBlock();
 
-	@property int Health();
+	@property float Health();
 
 	@property bool IsAttacking();
 	@property bool IsBlocking();
